move replace_calls to goto-instrument
This moves replace_calls to goto-instrument, which is the only user of this
module.
