+++ src/goto-instrument/Makefile
@@ -55,6 +55,7 @@ SRC = accelerate/accelerate.cpp \
       race_check.cpp \
       reachability_slicer.cpp \
       remove_function.cpp \
+      replace_calls.cpp \
       rw_set.cpp \
       show_locations.cpp \
       skip_loops.cpp \
